ext4: Fix possible use-after-free in ext4_find_extent
syzbot reported a use-after-free Read in ext4_find_extent that is hit when
using a corrupted file system. The bug was reported on Android 5.15, but
using the same reproducer tiggers the bug on 6.2.0-rc1 as well.

Fix the use-after-free by checking the extent header magic. An alternative
would be to check the values of EXT4_{FIRST,LAST}_{EXTENT,INDEX} used in
ext4_ext_binsearch() and ext4_ext_binsearch_idx(), so that we make sure
that pointers returned by EXT4_{FIRST,LAST}_{EXTENT,INDEX} don't exceed the
bounds of the extent tree node. But this alternative will not squash the bug
for the cases where eh->eh_entries fit into eh->eh_max. We could also try to
check the sanity of the path, but costs more than checking just the header
magic, so stick to the header magic sanity check.

Link: https://syzkaller.appspot.com/bug?id=be6e90ce70987950e6deb3bac8418344ca8b96cd
Reported-by: syzbot+0827b4b52b5ebf65f219@syzkaller.appspotmail.com
Cc: stable@vger.kernel.org
Change-Id: I680047828b715a5a8c852d0210994985df39865b
Signed-off-by: Tudor Ambarus <tudor.ambarus@linaro.org>